Disable swiftcall test on windows: More brutal way to appease windows bots
The backtrace on the bot does not give me any indication what is wrong.
The test case interestingly passes in stage2 of the build.
I don't have a way of debugging this.

Disable the test on windows and hope if there is truly a bug in the code that
was causing we will eventually run into this on other platforms.

git-svn-id: https://llvm.org/svn/llvm-project/cfe/trunk@284174 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit 3c62be7773a0e8295a24393d2311b135406781a9)

apple-llvm-split-commit: ad1235b36d6e061322e1bf2b2132c27b0c4af6a3
apple-llvm-split-dir: clang/