[AST] Remove superfluous docblock change
Upstream Clang documents the `Message` parameter as simply `\param`,
while Swift's fork of Clang documents it as `\param[out]`. Although
`\param[out]` may be more correct, the change to the documentation
should be made in upstream Clang, not Swift Clang.

Furthermore, upstream Clang does not appear to have a strong convention
as to the use of `\param[out]`: of 4616 instances of `\param`, only 49
are denoted as `\param[out]`.

apple-llvm-split-commit: 75035e4a3a2a80296b0a868e1e6150a860a336fb
apple-llvm-split-dir: clang/