[asan] Set the darwin deployment target for the dead-strip test
ASan dead-strip support relies on a linker option that only exists
in 10.11 and later, so the LLVM instrumentation checks for the deployment
target. This test does not pass when clang is built to choose lower
deployment target by default but runs on newer host.

(Note, the REQUIRES: osx-ld64-live_support clause only checks the host
and not the target OS.)

Differential Revision: https://reviews.llvm.org/D26107

git-svn-id: https://llvm.org/svn/llvm-project/compiler-rt/trunk@285482 91177308-0d34-0410-b5e6-96231b3b80d8
apple-llvm-split-commit: 24f066fa9ef663af6bec2b122642849729cf6fb9
apple-llvm-split-dir: compiler-rt/