[sanitizer] Make atos stdin a non-tty pipe to make sure it's not stuck waiting for user input
On macOS, we often symbolicate using atos (when llvm-symbolizer is not found). The current way we invoke atos involves creating a pseudo-terminal to make sure atos doesn't buffer its output. This however also makes atos think that it's stdin is interactive and in some error situations it will ask the user to enter some input instead of just printing out an error message. For example, when Developer Mode isn't enabled on a machine, atos cannot examine processes, and it will ask the user to enter an administrator's password, which will make the sanitized process get stuck. This patch only connects the pseudo-terminal to the stdout of atos, and uses a regular pipe as its stdin.

Differential Revision: https://reviews.llvm.org/D27239



git-svn-id: https://llvm.org/svn/llvm-project/compiler-rt/trunk@288624 91177308-0d34-0410-b5e6-96231b3b80d8
apple-llvm-split-commit: 67e6b95151971c503d493e4e55e21d911af42a5a
apple-llvm-split-dir: compiler-rt/