[sanitizer] Make sure libmalloc doesn't remove the sanitizer zone from malloc_zones[0]
In certain OS versions, it was possible that libmalloc replaced the sanitizer zone from being the default zone (i.e. being in malloc_zones[0]). This patch introduces a failsafe that makes sure we always stay the default zone. No testcase for this, because this doesn't reproduce under normal circumstances.

Differential Revision: https://reviews.llvm.org/D27083



git-svn-id: https://llvm.org/svn/llvm-project/compiler-rt/trunk@289376 91177308-0d34-0410-b5e6-96231b3b80d8
apple-llvm-split-commit: ba0e650e0da2165414363a452037902316477ff3
apple-llvm-split-dir: compiler-rt/