[CrashReproducer] Collect PCH included via -include-pch
Collect the necessary input PCH files.

Do not try to validate the AST before copying it out because if the
crash is in this path, we won't be able to collect it. Instead only
check if it's a file containg an AST.

rdar://problem/27913709

git-svn-id: https://llvm.org/svn/llvm-project/cfe/trunk@289460 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit 22d59b937301f26dde4bde2822452393f578962d)

apple-llvm-split-commit: d3ee0edb78c583c977faa986e0df86ed5269ceae
apple-llvm-split-dir: clang/