[APINotes] Remove VersionedInfoRole.
When API notes were first added, they would defer to the header when
something was explicitly specified---for example, if the API notes had
one set of nullability annotations from before the header was
audited. This behavior didn't make sense for versioned API notes,
however, where annotations needed for Swift 3 compatibility were
intended to override the unversioned information in both the API notes
and the header itself.

Now that the versioning of API notes is well-established, simplify
things by having the API notes always "win"; that is, an annotation
in API notes will always override what's in the headers.

apple-llvm-split-commit: b0755dd83ebfd484b843b3f63d87161af8c169e0
apple-llvm-split-dir: clang/